♻️ refactor: 유틸 통합 + config 최소화 및 파일 정리
pages, widget으로 분리된 유틸 통합
